feat: add SignRequest page
